requirements.txt
h5py==2.10.0
jupyter>=1.0.0,<2
jupyter_contrib_nbextensions==0.5.1
matplotlib==2.2.2
git+git://github.com/ionpath/mibilib@v1.2.6
netCDF4==1.5.3
numpy==1.16.3
pandas==0.24.2
scikit-image==0.16.2
scikit-learn==0.21.2
scipy==1.2.1
seaborn==0.10.1
statsmodels==0.11.1
tables==3.6.1
umap-learn==0.4.6
xarray==0.12.3
#Twisted~=20.3.0
#certifi
#pyOpenSSL
#service_identity
kiosk-client>=0.8.0

